refactor: move CLI args to new base-builder-cli crate
Create a new lightweight crate  in crates/builder/ to
separate CLI argument definitions from the op-rbuilder crate.

Moved to new crate:
- OpRbuilderArgs
- FlashblocksArgs
- TelemetryArgs
- GasLimiterArgs

The op-rbuilder crate now re-exports these types from base-builder-cli,
maintaining backwards compatibility for existing code.

Closes #471
