+++ crates/builder/base-builder-cli/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "base-builder-cli"
+description = "CLI argument definitions for the Base builder"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+# CLI
+clap.workspace = true
+
+# reth
+reth-optimism-node.workspace = true
+
+# misc
+anyhow = "1"
+shellexpand = "3.1"
