+++ crates/builder/op-rbuilder/src/flashblocks/mod.rs
@@ -2,7 +2,10 @@ use core::{convert::TryFrom, time::Duration};
 
 use reth_optimism_payload_builder::config::{OpDAConfig, OpGasLimitConfig};
 
-use crate::{args::OpRbuilderArgs, gas_limiter::args::GasLimiterArgs, tx_data_store::TxDataStore};
+use crate::{
+    args::{GasLimiterArgs, OpRbuilderArgs},
+    tx_data_store::TxDataStore,
+};
 
 pub(crate) mod best_txs;
 pub(crate) mod config;
